Fixed UpdatePings test patches and change file name
